+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.2 (13.01.2015)
+
+Features:
+
+* added **!db** - short PreDB stats
+* added **!day/!today, !week, !month, !year** - stats by time period
+* added size transformer - transforms MB size in better readable format (KB, GB, TB, PB)
+
 ## 1.1 (11.01.2015)
 
 Features: